refactor: Consolidate WorkOS authentication into single module
- Consolidated 4 fragmented auth files into convex/node/auth.ts
- Added comprehensive JSDoc documentation (144 lines)
- Deleted redundant files: types.ts, workos-auth.ts, middleware.ts, README_AUTH.md
- Maintained all existing function interfaces for zero breaking changes
- Validated with 26/26 integration tests passing

The authentication system is now properly organized with complete documentation
while eliminating code duplication and improving maintainability.
