--- apps/convex/lib/auth/types.ts
@@ -1,14 +1,0 @@
-/**
- * Authentication types for the Convex backend
- */
-
-export interface AuthenticatedUser {
-  id: string;
-  email: string;
-  customClaims: {
-    system_user?: string;
-    test_context?: string;
-    environment?: string;
-    permissions?: string[];
-  };
-}
